move listeners to buttons and circulars progress
