+++ app/src/main/java/skills/future/planer/ui/habit/HabitExtendedTotalAdapter.java
@@ -1,10 +1,7 @@
 package skills.future.planer.ui.habit;
 
 import android.annotation.SuppressLint;
-import android.app.AlertDialog;
 import android.content.Context;
-import android.content.Intent;
-import android.os.Bundle;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
@@ -12,7 +9,6 @@ import android.widget.ImageView;
 
 import androidx.annotation.NonNull;
 import androidx.fragment.app.Fragment;
-import androidx.lifecycle.ViewModelProvider;
 import androidx.recyclerview.widget.RecyclerView;
 
 import java.util.ArrayList;
@@ -20,12 +16,10 @@ import java.util.List;
 
 import skills.future.planer.R;
 import skills.future.planer.db.habit.HabitData;
-import skills.future.planer.db.habit.HabitViewModel;
 import skills.future.planer.ui.AnimateView;
 import skills.future.planer.ui.habit.view_holders.HabitExtendedViewHolder;
 
 public class HabitExtendedTotalAdapter extends RecyclerView.Adapter<HabitExtendedViewHolder> {
-
     private final LayoutInflater layoutInflater;
     private final Context context;
     private final Fragment fragment;
@@ -64,10 +58,6 @@ public class HabitExtendedTotalAdapter extends RecyclerView.Adapter<HabitExtende
         AnimateView.singleAnimation(itemView.findViewById(R.id.circularProgressIndicatorHabit), context, R.anim.scalezoom2);
         AnimateView.singleAnimation(itemView.findViewById(R.id.circularProgressIndicatorHabitDay), context, R.anim.scalezoom2);
         AnimateView.singleAnimation(itemView.findViewById(R.id.circularProgressIndicatorHabitDay), context, R.anim.scalezoom2);
-
-
-
-
         return itemView;
     }
 
@@ -78,37 +68,6 @@ public class HabitExtendedTotalAdapter extends RecyclerView.Adapter<HabitExtende
             HabitData current = habitsList.get(position);
             holder.setEveryThing(current);
         }
-
-        createListenerToEditButton(holder, position);
-        createListenerToTrashButton(holder, position);
-    }
-
-    private void createListenerToEditButton(@NonNull HabitExtendedViewHolder holder, int position) {
-        holder.itemView.findViewById(R.id.editImageHabit).setOnClickListener(e -> {
-            var intent = new Intent(fragment.getActivity(), HabitCreatorActivity.class);
-            var bundle = new Bundle();
-            bundle.putLong("habitToEditId", habitsList.get(position).getHabitId());
-            intent.putExtras(bundle);
-            fragment.requireActivity().startActivity(intent);
-            notifyItemChanged(position);
-        });
-    }
-
-    private void createListenerToTrashButton(@NonNull HabitExtendedViewHolder holder, int position) {
-        AlertDialog.Builder builder = new AlertDialog.Builder(context);
-
-        builder.setTitle("Confirm deletion");
-        builder.setMessage("Are you sure?");
-
-        builder.setPositiveButton("Yes", (dialog, which) -> {
-            new ViewModelProvider(fragment).get(HabitViewModel.class).delete(habitsList.get(position));
-            dialog.dismiss();
-        });
-
-        builder.setNegativeButton("No", (dialog, which) -> dialog.dismiss());
-
-        AlertDialog alert = builder.create();
-        holder.itemView.findViewById(R.id.trashImageViewHabit).setOnClickListener(e -> alert.show());
     }
 
     public long getItemId(int position) {
