+++ app/src/main/java/skills/future/planer/ui/habit/view_holders/HabitExtendedViewHolder.java
@@ -1,70 +1,64 @@
 package skills.future.planer.ui.habit.view_holders;
 
 import android.annotation.SuppressLint;
-import android.app.Activity;
+import android.app.AlertDialog;
 import android.content.Context;
+import android.content.Intent;
+import android.os.Bundle;
 import android.util.TypedValue;
 import android.view.View;
+import android.widget.ImageView;
 import android.widget.TextView;
 
 import androidx.activity.ComponentActivity;
-import androidx.core.content.ContextCompat;
-import androidx.fragment.app.Fragment;
 import androidx.lifecycle.ViewModelProvider;
 
 import com.google.android.material.chip.Chip;
 import com.google.android.material.chip.ChipGroup;
-import com.prolificinteractive.materialcalendarview.CalendarDay;
 
-import java.time.LocalDate;
-
-import antonkozyriatskyi.circularprogressindicator.CircularProgressIndicator;
 import lombok.Getter;
 import skills.future.planer.R;
 import skills.future.planer.db.goal.GoalData;
 import skills.future.planer.db.goal.GoalsViewModel;
 import skills.future.planer.db.habit.HabitData;
-import skills.future.planer.tools.DatesParser;
-import skills.future.planer.ui.day.views.habits.TextAdapter;
+import skills.future.planer.db.habit.HabitViewModel;
 import skills.future.planer.ui.goals.pager.recycler.MixedRecyclerElement;
-import skills.future.planer.ui.habit.HabitTextAdapter;
+import skills.future.planer.ui.habit.HabitCreatorActivity;
 
 @Getter
 public class HabitExtendedViewHolder extends HabitViewHolder {
-
-    private final CircularProgressIndicator circularProgressIndicatorHabit;
-    private final CircularProgressIndicator circularProgressIndicatorHabitDay;
     private final Context context;
     private final ChipGroup chipGroup;
     private final Chip goalChip;
     private final ComponentActivity activity;
     private final TextView goalLabel;
     private final View constraintLayout;
+    private final ImageView imageEdit;
+    private final ImageView imageTrash;
 
 
     public HabitExtendedViewHolder(View itemView, Context context, ComponentActivity activity) {
-        super(itemView);
-
-        circularProgressIndicatorHabit = itemView.findViewById(R.id.circularProgressIndicatorHabit);
-        circularProgressIndicatorHabitDay = itemView.findViewById(R.id.circularProgressIndicatorHabitDay);
+        super(itemView, context, activity);
         chipGroup = itemView.findViewById(R.id.chipGroupWeek);
         this.context = context;
-        this.activity=activity;
+        this.activity = activity;
         goalChip = itemView.findViewById(R.id.goalChip);
         goalLabel = itemView.findViewById(R.id.goalText);
         constraintLayout = itemView.findViewById(R.id.constraintLayout);
+        this.imageEdit = itemView.findViewById(R.id.editImageHabit);
+        this.imageTrash = itemView.findViewById(R.id.trashImageViewHabit);
     }
 
 
     @SuppressLint("ResourceAsColor")
     @Override
     public void setEveryThing(MixedRecyclerElement element) {
+        super.setEveryThing(element);
         if (element instanceof HabitData habitData) {
-            title.setText(habitData.getTitle());
             setUpChipGroup(habitData);
-            setUpCircularProgressIndicatorHabit(habitData);
-            setUpCircularProgressIndicatorOfDays(habitData);
             setUpGoalChip(habitData);
+            createListenerToEditButton(habitData);
+            createListenerToTrashButton(habitData);
         }
     }
 
@@ -81,84 +75,50 @@ public class HabitExtendedViewHolder extends HabitViewHolder {
         }
     }
 
-    private void setUpCircularProgressIndicatorHabit(HabitData habitData) {
-        circularProgressIndicatorHabit.setMaxProgress(100);
-
-        double currentProgress = countCurrentProgress(habitData);
-
-        circularProgressIndicatorHabit.setCurrentProgress(currentProgress);
-
-        circularProgressIndicatorHabit.setProgressTextAdapter(new TextAdapter());
-        if (circularProgressIndicatorHabit.getProgress() <= 40)
-            circularProgressIndicatorHabit.setProgressColor(ContextCompat.getColor(context, R.color.bad));
-        else if (circularProgressIndicatorHabit.getProgress() <= 75)
-            circularProgressIndicatorHabit.setProgressColor(ContextCompat.getColor(context, R.color.mid));
-        else
-            circularProgressIndicatorHabit.setProgressColor(ContextCompat.getColor(context, R.color.good));
-    }
-
-    /**
-     * Counts the progression for the habit from the start date to today
-     */
-    private double countCurrentProgress(HabitData habitData) {
-        LocalDate today = LocalDate.now();
-
-        LocalDate beginDate = DatesParser.toLocalDate(habitData.getBeginCalendarDay());
-
-        double days = 0;
-        double doneHabitDays = 0;
-
-        for (LocalDate i = beginDate; i.isBefore(today) || i.isEqual(today); i = i.plusDays(1)) {
-            if (habitData.isDayOfWeekChecked(i)) {
-                days++;
-                if (habitData.isHabitDone(DatesParser.toCalendarDay(i)))
-                    doneHabitDays++;
-            }
-        }
-        return (doneHabitDays / days) * 100;
-    }
-
-    private void setUpCircularProgressIndicatorOfDays(HabitData habitData) {
-        circularProgressIndicatorHabitDay.setMaxProgress(habitData.getHabitDuration().getDaysNumber());
-        if (CalendarDay.today().isAfter(habitData.getBeginCalendarDay()))
-            circularProgressIndicatorHabitDay
-                    .setCurrentProgress(DatesParser.countDifferenceBetweenDays(
-                            habitData.getBeginCalendarDay(), CalendarDay.today()) + 1);
-        else if (CalendarDay.today().isBefore(habitData.getBeginCalendarDay()))
-            circularProgressIndicatorHabitDay.setCurrentProgress(0);
-        else
-            circularProgressIndicatorHabitDay.setCurrentProgress(1);
-
-        HabitTextAdapter habitTextAdapter = new HabitTextAdapter();
-        habitTextAdapter.setMaxProgress(circularProgressIndicatorHabitDay.getMaxProgress());
-        circularProgressIndicatorHabitDay.setProgressTextAdapter(habitTextAdapter);
-
-        if (circularProgressIndicatorHabitDay.getProgress() / circularProgressIndicatorHabitDay.getMaxProgress() <= 0.4)
-            circularProgressIndicatorHabitDay.setProgressColor(ContextCompat.getColor(context, R.color.bad));
-        else if (circularProgressIndicatorHabitDay.getProgress() / circularProgressIndicatorHabitDay.getMaxProgress() <= 75)
-            circularProgressIndicatorHabitDay.setProgressColor(ContextCompat.getColor(context, R.color.mid));
-        else
-            circularProgressIndicatorHabitDay.setProgressColor(ContextCompat.getColor(context, R.color.good));
-    }
-
-    private void setUpGoalChip(HabitData habitData){
+    private void setUpGoalChip(HabitData habitData) {
         GoalsViewModel goalsViewModel = new ViewModelProvider(activity).get(GoalsViewModel.class);
         GoalData goal = goalsViewModel.findById(habitData.getForeignKeyToGoal());
-        if (goal != null){
+        if (goal != null) {
             goalChip.setVisibility(View.VISIBLE);
             goalLabel.setVisibility(View.VISIBLE);
-            int px = (int) TypedValue.applyDimension(TypedValue.COMPLEX_UNIT_SP, 300, activity.getResources().getDisplayMetrics());
-            constraintLayout.getLayoutParams().height=px;
+            constraintLayout.getLayoutParams().height = (int) TypedValue.applyDimension(TypedValue.COMPLEX_UNIT_SP, 300, activity.getResources().getDisplayMetrics());
             String goalText = goal.getTitle();
-            if(goalText.length()>30){
-                goalText = goalText.substring(0,27)+"...";
+            if (goalText.length() > 30) {
+                goalText = goalText.substring(0, 27) + "...";
             }
             goalChip.setText(goalText);
-        }else {
+        } else {
             goalChip.setVisibility(View.INVISIBLE);
             goalLabel.setVisibility(View.INVISIBLE);
-            int px = (int) TypedValue.applyDimension(TypedValue.COMPLEX_UNIT_SP, 220, activity.getResources().getDisplayMetrics());
-            constraintLayout.getLayoutParams().height=px;
+            constraintLayout.getLayoutParams().height = (int) TypedValue.applyDimension(TypedValue.COMPLEX_UNIT_SP, 220, activity.getResources().getDisplayMetrics());
         }
+    }
+
+    private void createListenerToEditButton(HabitData habitData) {
+        imageEdit.setOnClickListener(e -> {
+            var intent = new Intent(activity, HabitCreatorActivity.class);
+            var bundle = new Bundle();
+            bundle.putLong("habitToEditId", habitData.getHabitId());
+            intent.putExtras(bundle);
+            activity.startActivity(intent);
+            //notifyItemChanged(position);
+        });
+    }
+
+    private void createListenerToTrashButton(HabitData habitData) {
+        AlertDialog.Builder builder = new AlertDialog.Builder(context);
+
+        builder.setTitle("Confirm deletion");
+        builder.setMessage("Are you sure?");
+
+        builder.setPositiveButton("Yes", (dialog, which) -> {
+            new ViewModelProvider(activity).get(HabitViewModel.class).delete(habitData);
+            dialog.dismiss();
+        });
+
+        builder.setNegativeButton("No", (dialog, which) -> dialog.dismiss());
+
+        AlertDialog alert = builder.create();
+        imageTrash.setOnClickListener(e -> alert.show());
     }
 }
