+++ app/src/main/res/layout/fragment_habit_in_list.xml
@@ -51,7 +51,7 @@
 
             <antonkozyriatskyi.circularprogressindicator.CircularProgressIndicator
                 android:id="@+id/circularProgressIndicatorHabit"
-                android:layout_width="30dp"
+                android:layout_width="50dp"
                 android:layout_height="80dp"
                 android:layout_marginTop="2dp"
                 android:layout_marginEnd="2dp"
@@ -73,7 +73,7 @@
 
             <antonkozyriatskyi.circularprogressindicator.CircularProgressIndicator
                 android:id="@+id/circularProgressIndicatorHabitDay"
-                android:layout_width="30dp"
+                android:layout_width="50dp"
                 android:layout_height="80dp"
                 android:layout_marginTop="2dp"
                 android:layout_marginEnd="8dp"
